+++ src/test/testDirectives.ts
@@ -592,7 +592,7 @@ describe('@directives', () => {
             field.type = GraphQLString;
             field.resolve = async function (...args: any[]) {
               const date = await resolve.apply(this, args);
-              return formatDate(date, format);
+              return formatDate(date, format, true);
             };
           }
         }
@@ -601,7 +601,7 @@ describe('@directives', () => {
       resolvers: {
         Query: {
           today() {
-            return new Date(1519688273858);
+            return new Date(1519688273858).toUTCString();
           }
         }
       }
@@ -877,7 +877,7 @@ describe('@directives', () => {
 
     function checkErrors(
       expectedCount: number,
-      ...expectedNames: string[],
+      ...expectedNames: string[]
     ) {
       return function ({ errors = [], data }: {
         errors: any[],
@@ -1232,5 +1232,31 @@ describe('@directives', () => {
       peopleType.ofType,
       Human
     );
+  });
+
+  it('does not enforce query directive locations (issue #680)', () => {
+    const visited = new Set<GraphQLObjectType>();
+    const schema = makeExecutableSchema({
+      typeDefs: `
+      directive @hasScope(scope: [String]) on QUERY | FIELD
+
+      type Query @hasScope {
+        oyez: String
+      }`,
+
+      schemaDirectives: {
+        hasScope: class extends SchemaDirectiveVisitor {
+          public visitObject(object: GraphQLObjectType) {
+            assert.strictEqual(object.name, 'Query');
+            visited.add(object);
+          }
+        }
+      }
+    });
+
+    assert.strictEqual(visited.size, 1);
+    visited.forEach(object => {
+      assert.strictEqual(schema.getType('Query'), object);
+    });
   });
 });