+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### vNEXT
 
+* Fix timezone bug in test for @date directive [PR #686](https://github.com/apollographql/graphql-tools/pull/686)
+
 * Expose `defaultMergedResolver` from stitching [PR #685](https://github.com/apollographql/graphql-tools/pull/685)
 
 ### v2.23.0